backend/requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
psycopg
alembic
groq
celery
python-simhash
pgvector
httpx
orjson
prometheus-client
structlog
python-dotenv
httpx[http2]
selectolax
apscheduler
pymupdf
pdfminer.six
python-docx
python-multipart
Unidecode
PyPDF2